+++ flink-connector-clickhouse-2.0.0/src/test/java/org/apache/flink/connector/clickhouse/sink/convertor/DateTimePOJOConvertor.java
@@ -2,6 +2,7 @@ package org.apache.flink.connector.clickhouse.sink.convertor;
 
 import com.clickhouse.data.ClickHouseDataType;
 import com.clickhouse.utils.Serialize;
+import com.clickhouse.utils.writer.DataWriter;
 import org.apache.flink.connector.clickhouse.convertor.POJOConvertor;
 import org.apache.flink.connector.clickhouse.sink.pojo.DateTimePOJO;
 
@@ -11,14 +12,18 @@ import java.time.ZoneOffset;
 
 public class DateTimePOJOConvertor extends POJOConvertor<DateTimePOJO> {
 
+    public DateTimePOJOConvertor(boolean hasDefaults) {
+        super(hasDefaults);
+    }
+
     @Override
-    public void instrument(OutputStream out, DateTimePOJO input) throws IOException {
-        Serialize.writeString(out, input.id, false, false, ClickHouseDataType.String, false, "id");
+    public void instrument(DataWriter dataWriter, DateTimePOJO input) throws IOException {
+        dataWriter.writeString(input.id, false, ClickHouseDataType.String, false, "id");
         if (input.dataType.equals(DateTimePOJO.DataType.DATETIME64)) {
-            Serialize.writeTimeDate64(out, input.createdAt.atZone(ZoneOffset.UTC), false, false, ClickHouseDataType.DateTime64, false, "createdAt", input.precision);
+            dataWriter.writeTimeDate64(input.createdAt.atZone(ZoneOffset.UTC),false, ClickHouseDataType.DateTime64, false, "createdAt", input.precision);
         } else {
-            Serialize.writeTimeDate(out, input.createdAt.atZone(ZoneOffset.UTC), false, false, ClickHouseDataType.DateTime, false, "createdAt");
+            dataWriter.writeTimeDate(input.createdAt.atZone(ZoneOffset.UTC), false, ClickHouseDataType.DateTime, false, "createdAt");
         }
-        Serialize.writeInt32(out, input.numLogins, false, false, ClickHouseDataType.Int32, false, "numLogins");
+        dataWriter.writeInt32(input.numLogins, false, ClickHouseDataType.Int32, false, "numLogins");
     }
 }