+++ flink-connector-clickhouse-base/src/main/java/org/apache/flink/connector/clickhouse/convertor/POJOConvertor.java
@@ -1,16 +1,25 @@
 package org.apache.flink.connector.clickhouse.convertor;
 
+import com.clickhouse.utils.writer.DataWriter;
+
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.OutputStream;
 import java.io.Serializable;
 
 public abstract class POJOConvertor<InputT> implements Serializable {
-    public abstract void instrument(OutputStream out, InputT input) throws IOException;
+    private final boolean schemaHasDefaults;
+
+    public POJOConvertor(boolean schemaHasDefaults) {
+        this.schemaHasDefaults = schemaHasDefaults;
+    }
+
+    public abstract void instrument(DataWriter dataWriter, InputT input) throws IOException;
 
     public byte[] convert(InputT input) throws IOException {
         ByteArrayOutputStream out = new ByteArrayOutputStream();
-        instrument(out, input);
+        DataWriter dataWriter = DataWriter.of(schemaHasDefaults, out);
+        instrument(dataWriter, input);
         return out.toByteArray();
     }
+
 }